Removed the networking configuration for Spark, created a master/worker setup.
The Spark master was connected with the mongoDB via a docker style network. This caused issues
between the Spark master and worker, see the notes at
https://docs.docker.com/compose/compose-file/#links for more information.

Before, the Spark master ran next to the worker process. This is against the entire concept of
docker.Now, the master and worker run in a separate container. More workers can be added by
extending the docker compose file.